% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDescriptives.R
\name{combine_imputed_descriptives}
\alias{combine_imputed_descriptives}
\title{A function to calculate mean, SD and CIs over imputed dataset descriptives}
\usage{
combine_imputed_descriptives(dataset)
}
\arguments{
\item{dataset}{data analyses are run on - each column should be a variable of interest and each row an imputed dataset}
}
\value{
A data frame containing mean, SD and CIs for all imputed datasets
}
\description{
A function to calculate mean, SD and CIs over imputed dataset descriptives
}
\examples{
combine_imputed_descriptives(dataset)
}
