% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelComparison.R
\name{get_stats_on_fit_from_MI}
\alias{get_stats_on_fit_from_MI}
\title{A function to calculate mean, standard deviation and CIs for model fit stats across imputed datasets}
\usage{
get_stats_on_fit_from_MI(x, y)
}
\arguments{
\item{x}{A mids object}

\item{y}{Model to be evaluated as a string}
}
\value{
A numeric vector containing the weights mean, SD & upper and lower values of 95% interval for deviance and AIC
}
\description{
A function to calculate mean, standard deviation and CIs for model fit stats across imputed datasets
}
\examples{
get_stats_on_fit_from_MI(mi.res,"age5_standardised ~ gender")
}
