% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDescriptives.R
\name{get_weighted_descriptives}
\alias{get_weighted_descriptives}
\title{A function to create a mean, standard deviation and CIs for sample weighted data}
\usage{
get_weighted_descriptives(x, y)
}
\arguments{
\item{x}{A vector of values to be summarised. No NAs.}

\item{y}{A vector of weights for each of the values in x}
}
\value{
A numeric vector containing the weights mean, SD & upper and lower values of 95% interval
}
\description{
A function to create a mean, standard deviation and CIs for sample weighted data
}
\examples{
get_weighted_descriptives(languageability,sampleweights)
}
