% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDescriptives.R
\name{get_weighted_proportions}
\alias{get_weighted_proportions}
\title{A function to calculate group proportions for categorical variables in sample weighted data}
\usage{
get_weighted_proportions(x, y)
}
\arguments{
\item{x}{A vector of the factor for which proportions to be calculated.}

\item{y}{A vector of weights for each of the values in x}
}
\value{
A numeric vector containing the proportions for each level of factor
}
\description{
A function to calculate group proportions for categorical variables in sample weighted data
}
\examples{
get_weighted_proportions(highested,sampleweights)
}
