% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelComparison.R
\name{pool.compare.lrm}
\alias{pool.compare.lrm}
\title{A function to compare lrm models estimated over imputed datasets}
\usage{
pool.compare.lrm(fit1,fit0,data,method = c("wald", "likelihood"))
}
\arguments{
\item{method}{method used for comparison - wald or likelihood}

\item{model1}{}

\item{model0}{}

\item{dataset}{dataset analyses are run on}
}
\value{
A list containing Dm, dfs and pvalue
}
\description{
A function to compare lrm models estimated over imputed datasets
}
\examples{
Mh1_sampleweight_lrm <- with(imputed_mcs2, lrm(recoded_mh_d1 ~ standardised_bw + standardised_gest + sex_combine + ethnic_combine + new_langcombine + highested + standardised_wealth, weights = weight_combine))
Mh2_sampleweight_lrm <- with(imputed_mcs2, lrm(recoded_mh_d1 ~ standardised_bw + standardised_gest + sex_combine + ethnic_combine + new_langcombine + highested + standardised_wealth + teen_mum2 +  ms_combine + standardised_maternal_mh + standardised_conduct + standardised_emotional,weights = weight_combine))
pool.compare.lrm(Mh2_sampleweight_lrm,Mh1_sampleweight_lrm,imputed_mcs2)
}
