% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pooling.R
\name{pool.lrm}
\alias{pool.lrm}
\title{A function to pool model output from across multiple imputed datasets}
\usage{
pool.lrm(models,totalsamplesize)
}
\arguments{
\item{models}{Fitted models}

\item{totalsamplesize}{samples size for datasets used}
}
\value{
pooled coefficients etc for models
}
\description{
A function to pool model output from across multiple imputed datasets
}
\examples{
Mh1_sampleweight_lrm <- with(imputed_mcs2, lrm(recoded_mh_d1 ~ standardised_bw + standardised_gest + sex_combine + ethnic_combine + new_langcombine + highested + standardised_wealth, weights = weight_combine))
pool.lrm(Mh1_sampleweight_lrm,length(imputed_mcs2$data[,1]))
}
